# src/docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Library Rent Book API",
        "description": "API for managing library book rentals.",
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "operationId": "users",
                "parameters": [
                    {
                        "name": "user_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "book_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list matching users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersBody"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/create": {
            "post": {
                "tags": [
                    "user"
                ],
                "operationId": "create_user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "user created succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreatedUserBody"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/rent": {
            "post": {
                "tags": [
                    "user"
                ],
                "operationId": "rent_book",
                "parameters": [
                    {
                        "name": "nation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentBook"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "book rented succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RentedBookBody"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{nation_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "operationId": "get_user",
                "parameters": [
                    {
                        "name": "nation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserBody"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreatedUserBody": {
                "type": "object",
                "description": "Represents the body of a response when a user created.",
                "required": [
                    "info",
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "info": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
...
            "GetUserBody": {
                "type": "object",
                "description": "Represents the body of a response containing a user's history.",
                "required": [
                    "user"
                ],
                "properties": {
                    "user": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserHistoryRow"
                        }
                    }
                }
            },
            "RentBook": {
                "type": "object",
                "description": "Represents a book to be rented.",
                "required": [
                    "book_name",
                    "due_date"
                ],
                "properties": {
                    "book_name": {
                        "type": "string"
                    },
                    "due_date": {
                        "type": "string"
                    }
                }
            },
            "RentedBookBody": {
                "type": "object",
                "description": "Represents the body of a response when a user rents a book.",
                "required": [
                    "message",
                    "info"
                ],
                "properties": {
                    "info": {
                        "$ref": "#/components/schemas/UserRentBook"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "User": {
                "type": "object",
                "description": "Represents a user.",
                "required": [
                    "nation_id",
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "nation_id": {
                        "type": "string"
                    }
                }
            },
            "UserHistoryRow": {
                "type": "object",
                "description": "Represents a row in the user's rental history.",
                "required": [
                    "name",
                    "nation_id",
                    "book_name",
                    "due_date"
                ],
                "properties": {
                    "book_name": {
                        "type": "string"
                    },
                    "due_date": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "nation_id": {
                        "type": "string"
                    }
                }
            },
            "UserRentBook": {
                "type": "object",
                "description": "Represents a book rental by a user.",
                "required": [
                    "nation_id",
                    "book_name",
                    "due_date"
                ],
                "properties": {
                    "book_name": {
                        "type": "string"
                    },
                    "due_date": {
                        "type": "string"
                    },
                    "nation_id": {
                        "type": "string"
                    }
                }
            },
            "UserRow": {
                "type": "object",
                "description": "Represents a row in the user table.",
                "required": [
                    "nation_id",
                    "user_name",
                    "book_name"
                ],
                "properties": {
                    "book_name": {
                        "type": "string"
                    },
                    "nation_id": {
                        "type": "string"
                    },
                    "user_name": {
                        "type": "string"
                    }
                }
            },
            "UsersBody": {
                "type": "object",
                "description": "Represents the body of a response containing multiple users.",
                "required": [
                    "users"
                ],
                "properties": {
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserRow"
                        }
                    }
                }
            }
        }
    }
}